refactor to extract tlc certificates handling
